+++ CHANGELOG.md
@@ -1,4 +1,10 @@
 # java-mac Cookbook Change Log
 
+##1.1.0
+* Move cookie to node attribute
+* Fix comment for adding java_home to profile
+* Update supermarket source
+* Add missing dependency on line cookbook
+
 ##1.0.1
 * Bump to JDK/JRE 7u60